feat: 支持通过接口转接会话

+++ client.go
@@ -8,7 +8,7 @@ import (
 
 // BaseModel 基础数据
 type BaseModel struct {
-	ErrCode int    `json:"errcode"` // 出错返回码，为0表示成功，非0表示调用失败
+	ErrCode int64    `json:"errcode"` // 出错返回码，为0表示成功，非0表示调用失败
 	ErrMsg  string `json:"errmsg"`  // 返回码提示语
 }
